refactor: 폴더 및 파일 이름과 경로 수정
